chore: create `.gitattributes` for consistent EOL handling (#120)
This pull request includes changes to ensure consistent end-of-line
(EOL) settings across text files and updates to the sync-client
configuration to include the `.gitattributes` file.

Changes to EOL settings:

*
[`.gitattributes`](diffhunk://#diff-618cd5b83d62060ba3d027e314a21ceaf75d36067ff820db126642944145393eR1-R4):
Added configuration to ensure consistent EOL (LF) for all text files.

Updates to sync-client configuration:

*
[`.github/sync-client.yml`](diffhunk://#diff-93bc202766315b6269beef308a6ad30ed3e86938ddbfa31b49e030f2263695f1R52-R53):
Added `.gitattributes` to the list of files to be synchronized.