refactor(*): rename `src/types.js` to `src/types/index.js` (#175)
This pull request includes several changes to the `package.json` and
`src/textlint-rule-allowed-uris.js` files to enhance testing
capabilities and provide important documentation.

Testing enhancements:

*
[`package.json`](diffhunk://#diff-7ae45ad102eab3b6d7e7896acd08c427a9b25b346470d7bc6507b6481575d519R43):
Added a new script for integration tests that runs multiple textlint
commands.

Documentation updates:

*
[`src/textlint-rule-allowed-uris.js`](diffhunk://#diff-7da72e1e0d5c49a43b94f9d928d3157918c0bdcc3121c5ed14e0bc459b3f26f9R1-R6):
Added a file-level comment indicating the importance of renaming the
file to `index.js` for it to function correctly as a RULE ID.