chore(*): install `textlint` as deps-dev and delete unnecessary files (#190)
This pull request includes changes to the `package.json` file and the
removal of unnecessary files related to the
`textlint-rule-no-doubled-space` package. The most important changes are
listed below:

### Dependency Updates:
*
[`package.json`](diffhunk://#diff-7ae45ad102eab3b6d7e7896acd08c427a9b25b346470d7bc6507b6481575d519R74):
Added `textlint` and `textlint-tester` to the development dependencies.

### Cleanup:
*
[`packages/textlint-rule-no-doubled-space/README.md`](diffhunk://#diff-39a2c3cac0ff22b0b98b0713f700fd71343b0b98499d0b4146c2dc6857a774abL1-L3):
Removed the README file content for `textlint-rule-no-multiple-spaces`.
*
[`packages/textlint-rule-no-doubled-space/VScode.code-workspace`](diffhunk://#diff-df30106924ef9a82fe842e13a554a4d8b113de149934ea81c97e4f5ad16be876L1-L7):
Removed the VS Code workspace configuration file.