docs(*): create `SECURITY.md` (#193)
This pull request includes several changes to the repository, focusing
on configuration updates, documentation improvements, and a crucial code
comment correction. The most important changes include the addition of a
security policy, updates to the configuration sync file, and
modifications to the changelog and a critical comment in the code.

### Documentation Improvements:
*
[`SECURITY.md`](diffhunk://#diff-f6ed156e4bf5c791680662464b94ea5d753f219ee816b385f67870e2c0d7d4c7R1-R35):
Added a comprehensive security policy detailing how to report
vulnerabilities, the response process, best practices, and supported
versions.
*
[`CHANGELOG.md`](diffhunk://#diff-06572a96a58dc510037d5efa622f9bec8519bc1beab13c9f251e97e657a9d4edL1-R1):
Updated the header to "Change Log" and included links to releases and
tags in GitHub.

### Configuration Updates:
*
[`.github/sync-client.yml`](diffhunk://#diff-93bc202766315b6269beef308a6ad30ed3e86938ddbfa31b49e030f2263695f1R64-R65):
Added the `SECURITY.md` file to the sync configuration to ensure it is
included in the repository's configurations.

### Code Comment Correction:
*
[`src/textlint-rule-allowed-uris.js`](diffhunk://#diff-7da72e1e0d5c49a43b94f9d928d3157918c0bdcc3121c5ed14e0bc459b3f26f9L4-R4):
Corrected a critical comment to indicate that the filename must not be
changed to `index.js` as it is used as a RULE ID.