docs(*): update `README.md` and `config.yml` (#195)
This pull request includes various updates and improvements to the
`README.md` file and the `config.yml` file for better clarity and
consistency. The most important changes include fixing typos, updating
URLs, and improving instructions.

Improvements and fixes:

*
[`README.md`](diffhunk://#diff-b335630551682c19a781afebcf4d07bf978fb1f8ac04c6bf87428ed5106870f5L27-R27):
Fixed typos and improved clarity in several sections, including changing
"codes" to "code" and updating commit message instructions.
[[1]](diffhunk://#diff-b335630551682c19a781afebcf4d07bf978fb1f8ac04c6bf87428ed5106870f5L27-R27)
[[2]](diffhunk://#diff-b335630551682c19a781afebcf4d07bf978fb1f8ac04c6bf87428ed5106870f5L410-R410)
[[3]](diffhunk://#diff-b335630551682c19a781afebcf4d07bf978fb1f8ac04c6bf87428ed5106870f5L434-R434)
[[4]](diffhunk://#diff-b335630551682c19a781afebcf4d07bf978fb1f8ac04c6bf87428ed5106870f5L445-R450)
*
[`README.md`](diffhunk://#diff-b335630551682c19a781afebcf4d07bf978fb1f8ac04c6bf87428ed5106870f5L472-R472):
Added a note about the support for `LinkReference`, `ImageReference`,
and `Definition` types since `v14.5.0`.

URL updates:

*
[`.github/ISSUE_TEMPLATE/config.yml`](diffhunk://#diff-1c0d972ee49103af56fd608a77a28e1eb12f6908f263f0f183d46868fdcd8ea2L10-R10):
Updated the URL for the AST Tree for textlint to a more specific link.
*
[`README.md`](diffhunk://#diff-b335630551682c19a781afebcf4d07bf978fb1f8ac04c6bf87428ed5106870f5L27-R27):
Updated the link to the AST Tree of examples to the new URL format.