ci(*): create `release.yml` (#199)
This pull request includes changes to the GitHub Actions workflows to
add a new release workflow and update the sync-client configuration. The
most important changes include the addition of a new release workflow
configuration file and the update to the sync-client configuration file
to include the release workflow.

Updates to GitHub Actions workflows:

*
[`.github/workflows/release.yml`](diffhunk://#diff-87db21a973eed4fef5f32b267aa60fcee5cbdf03c67fafdc2a9b553bb0b15f34R1-R32):
Added a new release workflow that triggers on pull request closure,
checks specific conditions, and publishes a release using GitHub CLI.
*
[`.github/sync-client.yml`](diffhunk://#diff-93bc202766315b6269beef308a6ad30ed3e86938ddbfa31b49e030f2263695f1R18-R19):
Updated the sync-client configuration to include the new release
workflow.